+++ src/app/login/sing-in/sing-in.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-sing-in',
+  templateUrl: './sing-in.component.html',
+  styleUrls: ['./sing-in.component.scss']
+})
+export class SinginComponent {
+
+}
