+++ src/app/login/sing-up/sing-up.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-sing-up',
+  templateUrl: './sing-up.component.html',
+  styleUrls: ['./sing-up.component.scss']
+})
+export class SingupComponent {
+
+}
